+++ classes/Latte/MyCustomFilters.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace WorkOfStan\mycmsprojectnamespace\Latte;
+namespace WorkOfStan\MyCMS\Latte;
 
 use GodsDev\Tools\Tools;
 use Webmozart\Assert\Assert;
@@ -12,7 +12,7 @@ use WorkOfStan\MyCMS\MyCMS;
  * Custom project-specific filters for Latte.
  * (Last MyCMS/dist revision: 2022-02-04, v0.4.5)
  */
-class CustomFilters
+class MyCustomFilters
 {
     use \Nette\SmartObject;